Refactor(skill): Add skill system
