+++ src/qwenpaw/agents/skill_system/__init__.py
@@ -1,0 +1,40 @@
+# -*- coding: utf-8 -*-
+"""Skill system exports."""
+
+from .models import (
+    SkillConflictError,
+    SkillInfo,
+)
+from .pool_service import SkillPoolService
+from .registry import (
+    apply_skill_config_env_overrides,
+    ensure_skill_pool_initialized,
+    ensure_skills_initialized,
+    reconcile_pool_manifest,
+    reconcile_workspace_manifest,
+    resolve_effective_skills,
+)
+from .store import (
+    get_skill_pool_dir,
+    get_workspace_skills_dir,
+    read_skill_manifest,
+    read_skill_pool_manifest,
+)
+from .workspace_service import SkillService
+
+__all__ = [
+    "SkillConflictError",
+    "SkillInfo",
+    "SkillPoolService",
+    "SkillService",
+    "apply_skill_config_env_overrides",
+    "ensure_skill_pool_initialized",
+    "ensure_skills_initialized",
+    "get_skill_pool_dir",
+    "get_workspace_skills_dir",
+    "read_skill_manifest",
+    "read_skill_pool_manifest",
+    "reconcile_pool_manifest",
+    "reconcile_workspace_manifest",
+    "resolve_effective_skills",
+]
