+++ src/qwenpaw/cli/skills_cmd.py
@@ -6,19 +6,19 @@ from pathlib import Path
 
 import click
 
-from ..agents.skills_manager import (
+from ..agents.skill_system import (
     SkillConflictError,
     SkillPoolService,
     SkillService,
-    _validate_skill_content,
     get_workspace_skills_dir,
-    list_workspaces,
     read_skill_pool_manifest,
     read_skill_manifest,
     reconcile_pool_manifest,
     reconcile_workspace_manifest,
 )
-from ..agents.skills_hub import (
+from ..agents.skill_system.registry import list_workspaces
+from ..agents.skill_system.store import _validate_skill_content
+from ..agents.skill_system.hub import (
     import_pool_skill_from_hub,
     install_skill_from_hub,
 )
